requirements.txt
bayesian-optimization==1.0.1
CairoSVG==2.3.0
chaospy==3.0.4
cma==2.6.0
graphviz==0.10.1
joblib==0.13.2
matplotlib==3.0.3
networkx==2.2
numpy==1.16.2
panda3d==1.10.1
pandas==0.24.1
prettytable==0.7.2
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
Shapely==1.6.4.post2
sobol-seq==0.1.2
solidpython==0.3.1
svgwrite==1.2.1
tqdm==4.31.1
trimesh==2.37.3
